app_server/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
Werkzeug==2.2.3
PyMySQL==1.0.3

